// src/generated/protobuf/shared.ts
export enum Kind {
  AUDIO = 0,
  VIDEO = 1,
  UNRECOGNIZED = -1,
}
// ...
export interface Receiver {
  kind: Kind;
  name: string;
  state: Receiver_State | undefined;
}
// ...
export interface Receiver_Source {
  peer: string;
  track: string;
}

export interface Receiver_Config {
  priority: number;
  maxSpatial: number;
  maxTemporal: number;
  minSpatial?: number | undefined;
  minTemporal?: number | undefined;
}

export interface Receiver_State {
  config: Receiver_Config | undefined;
  source?: Receiver_Source | undefined;
}
// ...
export const Receiver = {
  encode(
    message: Receiver,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.kind !== 0) {
      writer.uint32(8).int32(message.kind);
    }
    if (message.name !== '') {
      writer.uint32(18).string(message.name);
    }
    if (message.state !== undefined) {
      Receiver_State.encode(message.state, writer.uint32(26).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Receiver {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseReceiver();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.kind = reader.int32() as any;
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.name = reader.string();
          continue;
        case 3:
          if (tag !== 26) {
            break;
          }

          message.state = Receiver_State.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Receiver {
    return {
      kind: isSet(object.kind) ? kindFromJSON(object.kind) : 0,
      name: isSet(object.name) ? globalThis.String(object.name) : '',
      state: isSet(object.state)
        ? Receiver_State.fromJSON(object.state)
        : undefined,
    };
  },

  toJSON(message: Receiver): unknown {
    const obj: any = {};
    if (message.kind !== 0) {
      obj.kind = kindToJSON(message.kind);
    }
    if (message.name !== '') {
      obj.name = message.name;
    }
    if (message.state !== undefined) {
      obj.state = Receiver_State.toJSON(message.state);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Receiver>, I>>(base?: I): Receiver {
    return Receiver.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Receiver>, I>>(object: I): Receiver {
    const message = createBaseReceiver();
    message.kind = object.kind ?? 0;
    message.name = object.name ?? '';
    message.state =
      object.state !== undefined && object.state !== null
        ? Receiver_State.fromPartial(object.state)
        : undefined;
    return message;
  },
};

function createBaseReceiver_Source(): Receiver_Source {
  return { peer: '', track: '' };
}

export const Receiver_Source = {
  encode(
    message: Receiver_Source,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.peer !== '') {
      writer.uint32(10).string(message.peer);
    }
    if (message.track !== '') {
      writer.uint32(18).string(message.track);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Receiver_Source {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseReceiver_Source();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.peer = reader.string();
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.track = reader.string();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Receiver_Source {
    return {
      peer: isSet(object.peer) ? globalThis.String(object.peer) : '',
      track: isSet(object.track) ? globalThis.String(object.track) : '',
    };
  },

  toJSON(message: Receiver_Source): unknown {
    const obj: any = {};
    if (message.peer !== '') {
      obj.peer = message.peer;
    }
    if (message.track !== '') {
      obj.track = message.track;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Receiver_Source>, I>>(
    base?: I,
  ): Receiver_Source {
    return Receiver_Source.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Receiver_Source>, I>>(
    object: I,
  ): Receiver_Source {
    const message = createBaseReceiver_Source();
    message.peer = object.peer ?? '';
    message.track = object.track ?? '';
    return message;
  },
};

function createBaseReceiver_Config(): Receiver_Config {
  return {
    priority: 0,
    maxSpatial: 0,
    maxTemporal: 0,
    minSpatial: undefined,
    minTemporal: undefined,
  };
}

export const Receiver_Config = {
  encode(
    message: Receiver_Config,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.priority !== 0) {
      writer.uint32(8).uint32(message.priority);
    }
    if (message.maxSpatial !== 0) {
      writer.uint32(16).uint32(message.maxSpatial);
    }
    if (message.maxTemporal !== 0) {
      writer.uint32(24).uint32(message.maxTemporal);
    }
    if (message.minSpatial !== undefined) {
      writer.uint32(32).uint32(message.minSpatial);
    }
    if (message.minTemporal !== undefined) {
      writer.uint32(40).uint32(message.minTemporal);
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Receiver_Config {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseReceiver_Config();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 8) {
            break;
          }

          message.priority = reader.uint32();
          continue;
        case 2:
          if (tag !== 16) {
            break;
          }

          message.maxSpatial = reader.uint32();
          continue;
        case 3:
          if (tag !== 24) {
            break;
          }

          message.maxTemporal = reader.uint32();
          continue;
        case 4:
          if (tag !== 32) {
            break;
          }

          message.minSpatial = reader.uint32();
          continue;
        case 5:
          if (tag !== 40) {
            break;
          }

          message.minTemporal = reader.uint32();
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Receiver_Config {
    return {
      priority: isSet(object.priority) ? globalThis.Number(object.priority) : 0,
      maxSpatial: isSet(object.maxSpatial)
        ? globalThis.Number(object.maxSpatial)
        : 0,
      maxTemporal: isSet(object.maxTemporal)
        ? globalThis.Number(object.maxTemporal)
        : 0,
      minSpatial: isSet(object.minSpatial)
        ? globalThis.Number(object.minSpatial)
        : undefined,
      minTemporal: isSet(object.minTemporal)
        ? globalThis.Number(object.minTemporal)
        : undefined,
    };
  },

  toJSON(message: Receiver_Config): unknown {
    const obj: any = {};
    if (message.priority !== 0) {
      obj.priority = Math.round(message.priority);
    }
    if (message.maxSpatial !== 0) {
      obj.maxSpatial = Math.round(message.maxSpatial);
    }
    if (message.maxTemporal !== 0) {
      obj.maxTemporal = Math.round(message.maxTemporal);
    }
    if (message.minSpatial !== undefined) {
      obj.minSpatial = Math.round(message.minSpatial);
    }
    if (message.minTemporal !== undefined) {
      obj.minTemporal = Math.round(message.minTemporal);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Receiver_Config>, I>>(
    base?: I,
  ): Receiver_Config {
    return Receiver_Config.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Receiver_Config>, I>>(
    object: I,
  ): Receiver_Config {
    const message = createBaseReceiver_Config();
    message.priority = object.priority ?? 0;
    message.maxSpatial = object.maxSpatial ?? 0;
    message.maxTemporal = object.maxTemporal ?? 0;
    message.minSpatial = object.minSpatial ?? undefined;
    message.minTemporal = object.minTemporal ?? undefined;
    return message;
  },
};

function createBaseReceiver_State(): Receiver_State {
  return { config: undefined, source: undefined };
}

export const Receiver_State = {
  encode(
    message: Receiver_State,
    writer: _m0.Writer = _m0.Writer.create(),
  ): _m0.Writer {
    if (message.config !== undefined) {
      Receiver_Config.encode(message.config, writer.uint32(10).fork()).ldelim();
    }
    if (message.source !== undefined) {
      Receiver_Source.encode(message.source, writer.uint32(18).fork()).ldelim();
    }
    return writer;
  },

  decode(input: _m0.Reader | Uint8Array, length?: number): Receiver_State {
    const reader =
      input instanceof _m0.Reader ? input : _m0.Reader.create(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseReceiver_State();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          if (tag !== 10) {
            break;
          }

          message.config = Receiver_Config.decode(reader, reader.uint32());
          continue;
        case 2:
          if (tag !== 18) {
            break;
          }

          message.source = Receiver_Source.decode(reader, reader.uint32());
          continue;
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skipType(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): Receiver_State {
    return {
      config: isSet(object.config)
        ? Receiver_Config.fromJSON(object.config)
        : undefined,
      source: isSet(object.source)
        ? Receiver_Source.fromJSON(object.source)
        : undefined,
    };
  },

  toJSON(message: Receiver_State): unknown {
    const obj: any = {};
    if (message.config !== undefined) {
      obj.config = Receiver_Config.toJSON(message.config);
    }
    if (message.source !== undefined) {
      obj.source = Receiver_Source.toJSON(message.source);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Receiver_State>, I>>(
    base?: I,
  ): Receiver_State {
    return Receiver_State.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Receiver_State>, I>>(
    object: I,
  ): Receiver_State {
    const message = createBaseReceiver_State();
    message.config =
      object.config !== undefined && object.config !== null
        ? Receiver_Config.fromPartial(object.config)
        : undefined;
    message.source =
      object.source !== undefined && object.source !== null
        ? Receiver_Source.fromPartial(object.source)
        : undefined;
    return message;
  },
};
// ...
type Builtin =
  | Date
  | Function
  | Uint8Array
  | string
  | number
  | boolean
  | undefined;

export type DeepPartial<T> = T extends Builtin
  ? T
  : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
      ? ReadonlyArray<DeepPartial<U>>
      : T extends {}
        ? { [K in keyof T]?: DeepPartial<T[K]> }
        : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin
  ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & {
      [K in Exclude<keyof I, KeysOfUnion<P>>]: never;
    };